requirements.txt
pandas
sqlalchemy
psycopg2-binary
sklearn
matplotlib
seaborn
joblib
numpy